+++ dpsim-villas/examples/cxx/CMakeLists.txt
@@ -16,10 +16,8 @@ set(SHMEM_SOURCES
 	FileExample.cpp
 	MqttExample.cpp
 	SharedMemExample.cpp
-	CosimExample_ext_control.cpp
-	CosimExample_ext_control_base.cpp
 	SocketExample.cpp
-	SocketCosimExample.cpp
+	CosimSocketExample.cpp
 	#ShmemExample.cpp
 	#ShmemDistributedReference.cpp
 	#ShmemDistributedDirect.cpp
